+++ Assets/Prefabs/Cat.prefab
@@ -11,7 +11,8 @@ GameObject:
   - component: {fileID: 207835099897388149}
   - component: {fileID: 207835099897388148}
   - component: {fileID: 207835099897388146}
-  - component: {fileID: 207835099897388147}
+  - component: {fileID: -5605180671712718243}
+  - component: {fileID: 8577323250487324892}
   m_Layer: 0
   m_Name: Cat
   m_TagString: Draggable
@@ -75,7 +76,7 @@ SpriteRenderer:
   m_SortingLayerID: 208536839
   m_SortingLayer: 1
   m_SortingOrder: 0
-  m_Sprite: {fileID: 7482667652216324306, guid: 311925a002f4447b3a28927169b83ea6, type: 3}
+  m_Sprite: {fileID: -2413806693520163455, guid: 4adeaf2d2b1c7417e89f026626d8264f, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
@@ -112,7 +113,7 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!114 &207835099897388147
+--- !u!114 &-5605180671712718243
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -121,7 +122,19 @@ MonoBehaviour:
   m_GameObject: {fileID: 207835099897388151}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6e6dfa29988a82f4ba90d06d65b87b4e, type: 3}
+  m_Script: {fileID: 11500000, guid: 93ffff2e7e91146c6b0fad0f8bba9f96, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &8577323250487324892
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 207835099897388151}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d6a2c1b1a46434e48b76a686217be1af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   currentRoom: {fileID: 0}
