--- Assets/Scripts/Cat/CatColours.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 08a20dcea43964034b0c3ebce698d821
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
