+++ Assets/Scripts/Cat/CatProperties.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93ffff2e7e91146c6b0fad0f8bba9f96
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
