+++ Assets/Scripts/Cat/CurrentRoom.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6e6dfa29988a82f4ba90d06d65b87b4e
+guid: d6a2c1b1a46434e48b76a686217be1af
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
