+++ Assets/Scripts/Game Control/SceneControl.cs
@@ -5,7 +5,12 @@ using UnityEngine.SceneManagement;
 
 public class SceneControl : MonoBehaviour
 {
+    //show catlist
    public void ListScene() {  
         SceneManager.LoadScene("CatList");  
     }  
+    //show main game (keanna for now cause testing)
+    public void GameScene() {  
+        SceneManager.LoadScene("KeannaScene");  
+    }
 }
